requirements.txt
Django==4.2.1
django-admin-sortable2==2.1.8
django-environ==0.10.0
django-tinymce==3.6.1
Pillow==9.5.0
requests==2.30.0
